Menu/Entrees/VelociWrap.cs
/* VelociWrap.cs
 * Auther: Mike Ruckert
 */
using System.Collections.Generic;

namespace DinoDiner.Menu
{
    /// <summary>
    /// Represents the VelociWrap entree menu item.
    /// </summary>
    public class VelociWrap : Entree
    {
        /// <summary>
        /// Indicates whether dressing is included in the ingredients.
        /// </summary>
        private bool dressing = true;

        /// <summary>
        /// Indicates whether lettuce is included in the ingredients.
        /// </summary>
        private bool lettuce = true;

        /// <summary>
        /// Indicates whether parmesan cheese is included in the ingredients.
        /// </summary>
        private bool cheese = true;

        /// <summary>
        /// Returns the description of this order item
        /// </summary>
        public override string Description
        {
            get { return this.ToString(); }
        }

        /// <summary>
        /// Overrides the Ingredients property in Entree.
        /// </summary>        
        public override List<string> Ingredients
        {
            get
            {
                List<string> ingredients = new List<string>() { "Flour Tortilla", "Chicken Breast" };
                if (dressing) ingredients.Add("Ceasar Dressing");
                if (lettuce) ingredients.Add("Romaine Lettuce");
                if (cheese) ingredients.Add("Parmesan Cheese");
                return ingredients;
            }
        }

        /// <summary>
        /// Gets any special preparation instructions
        /// </summary>
        public override string[] Special
        {
            get
            {
                List<string> specials = new List<string>();
                if (!dressing) specials.Add("Hold Ceasar Dressing");
                if (!lettuce) specials.Add("Hold Lettuce");
                if (!cheese) specials.Add("Hold Cheese");
                return specials.ToArray();
            }
        }

        /// <summary>
        /// Class constructor.  Sets the Price and Calories.
        /// </summary>
        public VelociWrap()
        {
            this.Price = 6.86;
            this.Calories = 356;
        }

        /// <summary>
        /// Method to hold the dressing.
        /// </summary>
        public void HoldDressing()
        {
            this.dressing = false;
            NotifyOfPropertyChange("Special");
            NotifyOfPropertyChange("Ingredients");
        }

        /// <summary>
        /// Method to hold the lettuce.
        /// </summary>
        public void HoldLettuce()
        {
            this.lettuce = false;
            NotifyOfPropertyChange("Special");
            NotifyOfPropertyChange("Ingredients");
        }

        /// <summary>
        /// Method to hold the parmesan cheese.
        /// </summary>
        public void HoldCheese()
        {
            this.cheese = false;
            NotifyOfPropertyChange("Special");
            NotifyOfPropertyChange("Ingredients");
        }

        /// <summary>
        /// Overrides the default ToString()
        /// </summary>
        /// <returns></returns>
        public override string ToString()
        {
            return "Veloci-Wrap";
        }


    }
}
